Separated reflection utilities in `ReflectionUtils`
