+++ src/main/java/com/freya02/botcommands/internal/application/slash/autocomplete/AutocompletionHandlersBuilder.java
@@ -6,6 +6,7 @@ import com.freya02.botcommands.internal.Logging;
 import com.freya02.botcommands.internal.MethodParameters;
 import com.freya02.botcommands.internal.application.slash.SlashCommandInfo;
 import com.freya02.botcommands.internal.application.slash.SlashCommandParameter;
+import com.freya02.botcommands.internal.utils.ReflectionUtils;
 import com.freya02.botcommands.internal.utils.Utils;
 import net.dv8tion.jda.api.events.interaction.CommandAutoCompleteEvent;
 import org.slf4j.Logger;
@@ -26,7 +27,7 @@ public class AutocompletionHandlersBuilder {
 
 	public void processHandler(Object autocompleteHandler, Method method) {
 		try {
-			if (!Utils.hasFirstParameter(method, CommandAutoCompleteEvent.class))
+			if (!ReflectionUtils.hasFirstParameter(method, CommandAutoCompleteEvent.class))
 				throw new IllegalArgumentException("Autocompletion handler at " + Utils.formatMethodShort(method) + " must have a " + CommandAutoCompleteEvent.class.getSimpleName() + " event as first parameter");
 
 			final AutocompletionHandlerInfo handler = new AutocompletionHandlerInfo(context, autocompleteHandler, method);
